get rid of evaluator module, combine it with the expression module
