+++ src/lib/modules/expression/index.js
@@ -3,8 +3,9 @@
 const Promise = require('bluebird');
 const _ = require('underscore');
 const esprima = require('esprima');
+const vm = require('vm');
+const ObjectId = require('mongodb').ObjectId;
 
-const evaluator = require('lib/modules/evaluator');
 const keyValueUtils = require('src/lib/utils/keyValueUtils');
 
 class Expression {
@@ -15,17 +16,11 @@ class Expression {
 
       let astTokens = esprima.tokenize(expression);
 
-      let evalScope = {
-        db: {}
-      };
-
-      collections.forEach(collection => {
-        evalScope.db[collection.name] = collection;
-      });
+      let evalScope = _createEvalScope(collections);
 
       let startTime = process.hrtime();
 
-      evaluator.eval(expression, evalScope)
+      _eval(expression, evalScope)
         .then((result) => {
           let time = _getTime(startTime);
 
@@ -62,6 +57,18 @@ class Expression {
   }
 }
 
+function _createEvalScope(collections) {
+  let evalScope = {
+    db: {}
+  };
+
+  collections.forEach(collection => {
+    evalScope.db[collection.name] = collection;
+  });
+
+  return evalScope;
+}
+
 function _getMongoMethodName(astTokens) {
   //TODO: need to handle the case of 'db[Cars].find'
   if (!astTokens || astTokens.length < 4) return null;
@@ -87,6 +94,63 @@ function _getMongoQuery() {
 function _getTime(startTime) {
   let endTime = process.hrtime(startTime);
   return endTime[0], endTime[1] / 1000000;
+}
+
+/** @private */
+/**
+ * evaluate a JS expression
+ * @param {String} expression
+ * @param {Object} scope
+ */
+function _eval(expression, scope) {
+  return new Promise((resolve, reject) => {
+    if (!expression) return reject(new Error('evaluator - eval() - must pass an expression'));
+    if (!_.isString(expression)) return reject(new Error('evaluator - eval() - expression must be a string'));
+
+    var evalScope = {
+      ObjectId: ObjectId
+    };
+
+    if (scope && _.isObject(scope)) {
+      _.extend(evalScope, scope);
+    }
+
+    var options = {
+      displayErrors: false
+    };
+
+    // Evalutate the expression with the given scope
+    var script;
+    var result;
+
+    try {
+      script = new vm.Script(expression, options);
+    } catch (err) {
+      return reject(err);
+    }
+
+    //NOTE: ast could contain multiple expressions (top level nodes)
+
+    try {
+      result = script.runInNewContext(evalScope, options);
+    } catch (err) {
+      return reject(err);
+    }
+
+    if (_isPromise(result)) {
+      result.then(promiseResult => {
+        return resolve(promiseResult);
+      });
+
+      if (result.catch) result.catch(reject);
+    } else {
+      return resolve(result);
+    }
+  });
+}
+
+function _isPromise(func) {
+  return func && func.then && typeof(func.then) === 'function';
 }
 
 module.exports = new Expression();
